쿠키 삭제 함수 모듈화
